Deletion handlers no-op if the targets do not exist.

+++ delete.go
@@ -6,6 +6,7 @@ import (
     "encoding/json"
     "path/filepath"
     "time"
+    "errors"
 )
 
 func deleteProjectHandler(reqpath string, globals *globalConfiguration) error {
@@ -38,6 +39,9 @@ func deleteProjectHandler(reqpath string, globals *globalConfiguration) error {
     }
 
     project_dir := filepath.Join(globals.Registry, *(incoming.Project))
+    if _, err := os.Stat(project_dir); errors.Is(err, os.ErrNotExist) {
+        return nil
+    }
     err = os.RemoveAll(project_dir)
     if err != nil {
         return fmt.Errorf("failed to delete %s; %v", project_dir, err)
@@ -90,6 +94,9 @@ func deleteAssetHandler(reqpath string, globals *globalConfiguration) error {
     }
 
     project_dir := filepath.Join(globals.Registry, *(incoming.Project))
+    if _, err := os.Stat(project_dir); errors.Is(err, os.ErrNotExist) {
+        return nil
+    }
     err = globals.Locks.LockPath(project_dir, 1000 * time.Second)
     if err != nil {
         return fmt.Errorf("failed to lock project directory %q; %w", project_dir, err)
@@ -97,6 +104,9 @@ func deleteAssetHandler(reqpath string, globals *globalConfiguration) error {
     defer globals.Locks.UnlockPath(project_dir)
 
     asset_dir := filepath.Join(project_dir, *(incoming.Asset))
+    if _, err := os.Stat(asset_dir); errors.Is(err, os.ErrNotExist) {
+        return nil
+    }
     to_free, err := computeUsage(asset_dir, true)
     if err != nil {
         return fmt.Errorf("failed to compute usage for %s; %v", asset_dir, err)
@@ -176,6 +186,9 @@ func deleteVersionHandler(reqpath string, globals *globalConfiguration) error {
     // project even if we're mutating a single asset and (ii) we need to update
     // the usage anyway so we'd have to obtain this lock eventually.
     project_dir := filepath.Join(globals.Registry, *(incoming.Project))
+    if _, err := os.Stat(project_dir); errors.Is(err, os.ErrNotExist) {
+        return nil
+    }
     err = globals.Locks.LockPath(project_dir, 1000 * time.Second)
     if err != nil {
         return fmt.Errorf("failed to lock project directory %q; %w", project_dir, err)
@@ -183,7 +196,14 @@ func deleteVersionHandler(reqpath string, globals *globalConfiguration) error {
     defer globals.Locks.UnlockPath(project_dir)
 
     asset_dir := filepath.Join(project_dir, *(incoming.Asset))
+    if _, err := os.Stat(asset_dir); errors.Is(err, os.ErrNotExist) {
+        return nil
+    }
+
     version_dir := filepath.Join(asset_dir, *(incoming.Version))
+    if _, err := os.Stat(version_dir); errors.Is(err, os.ErrNotExist) {
+        return nil
+    }
     to_free, err := computeUsage(version_dir, true)
     if err != nil {
         return fmt.Errorf("failed to compute usage for %s; %v", version_dir, err)
